AstronomicalTwilightTracker
This code calculates the start and end times of astronomical twilight based on your location on Earth. By tracking when the sun’s altitude falls below -18 degrees, the code determines the period of astronomical twilight, during which the sky is dark enough for astronomical observations. Using np.where to identify the first and last instances where the sun’s altitude meets this criterion, the code outputs the twilight start and end times in ISO format. This tracker is useful for astronomers and night sky enthusiasts who need precise timing for optimal stargazing or astrophotography conditions. CAUTION! The returned time values will be in Universal Time (UTC) for your respective region, i.e., the time is correctly outputter, however in UTC! [Please make the relevant conversions using an online clock]